requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
python-docx
openai
